mocks for rhc connection dao moved into separate file (mock_daos.go -> mock_dao_rhc_connection.go)
